requirements.txt
mistralai>=1.0.0
Pillow>=10.0.0
piexif>=1.1.3
python-dotenv>=1.0.0
